refactor: Removed deprecated references to listener utility
refactor: ambersteel-base-actor-sheet: Moved some logic to actor-sheet

* This is an effort to ultimately deprecate the class entirely. The plan is to move away from the current pseudo-inheritance composition architecture and make better use of object-oriented approaches or composition.
